requirements.txt
curl_cffi>=0.5.0
psycopg2-binary>=2.9.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
